Move import from epics.core to core
