+++ src/ophyd_async/core/__init__.py
@@ -16,6 +16,16 @@ from ._detector import (
 )
 from ._device import Device, DeviceConnector, DeviceVector, init_devices
 from ._device_filler import DeviceFiller
+from ._enums import (
+    EnabledState,
+    EnabledStateCapitalised,
+    EnableState,
+    EnableStateCapitalised,
+    InState,
+    InStateCapitalised,
+    OnState,
+    OnStateCapitalised,
+)
 from ._flyer import FlyerController, FlyMotorInfo, StandardFlyer
 from ._hdf_dataset import HDFDatasetDescription, HDFDocumentComposer
 from ._log import config_ophyd_async_logging
@@ -209,4 +219,13 @@ __all__ = [
     # Back compat - delete before 1.0
     "ConfigSignal",
     "HintedSignal",
+    # Common enums
+    "EnabledState",
+    "EnabledStateCapitalised",
+    "EnableState",
+    "EnableStateCapitalised",
+    "InState",
+    "InStateCapitalised",
+    "OnState",
+    "OnStateCapitalised",
 ]
