+++ src/ophyd_async/epics/adaravis/_aravis_controller.py
@@ -1,8 +1,7 @@
 import asyncio
 
-from ophyd_async.core import DetectorTrigger, TriggerInfo
+from ophyd_async.core import DetectorTrigger, OnState, TriggerInfo
 from ophyd_async.epics import adcore
-from ophyd_async.epics.core._enums import OnState
 
 from ._aravis_io import AravisDriverIO, AravisTriggerSource
