+++ src/ophyd_async/epics/adcore/_core_logic.py
@@ -6,12 +6,12 @@ from ophyd_async.core import (
     AsyncStatus,
     DetectorController,
     DetectorTrigger,
+    EnableState,
     TriggerInfo,
     observe_value,
     set_and_wait_for_value,
 )
 from ophyd_async.epics.core import stop_busy_record
-from ophyd_async.epics.core._enums import EnableState
 
 from ._core_io import (
     ADBaseIO,
