+++ src/ophyd_async/epics/adcore/_core_writer.py
@@ -12,6 +12,7 @@ from event_model import (  # type: ignore
 from pydantic import PositiveInt
 
 from ophyd_async.core._detector import DetectorWriter
+from ophyd_async.core._enums import EnableState
 from ophyd_async.core._providers import DatasetDescriber, PathInfo, PathProvider
 from ophyd_async.core._signal import (
     observe_value,
@@ -20,7 +21,6 @@ from ophyd_async.core._signal import (
 from ophyd_async.core._status import AsyncStatus
 from ophyd_async.core._utils import DEFAULT_TIMEOUT, error_if_none
 from ophyd_async.epics.core import stop_busy_record
-from ophyd_async.epics.core._enums import EnableState
 
 # from ophyd_async.epics.adcore._core_logic import ADBaseDatasetDescriber
 from ._core_io import (
